editing die.fish

+++ die.fish
@@ -1,11 +1,10 @@
 function die
     if exists $argv
-       switch $argv[1]
-            case "-p"
-                   killport $argv[2]
-              case "-n"
-                foreach (pgrep "$argv[2]") "kill -9 _"
-        end
+       if isnumeric $argv[2]
+          killport $argv[2]
+       else
+           foreach (pgrep "$argv[2]") "kill -9 _"
+       end
     else
           xkill
     end
